--- app/pages/posts.vue
@@ -1,3 +1,0 @@
-<template>
-  <NuxtPage />
-</template>
