+++ content/index.md
@@ -1,4 +1,8 @@
-title: Blog
+---
+title: A Onetime Blog
 description: Discover best practices for secure information sharing, learn essential privacy tips, and stay up-to-date with the latest Onetime Secret features. Our blog is your trusted resource for mastering one-time, self-destructing message delivery.
 navigation.icon: i-heroicons-newspaper
 ---
+
+
+Here's page content
